+++ src/Composants/Footer.jsx
@@ -11,12 +11,6 @@ function Footer() {
                             <a href="/about" className=" dark:hover:text-violet-700">About</a>
                         </li>
                         <li className="mb-2">
-                            <a href="#" className="dark:hover:text-violet-700">Careers</a>
-                        </li>
-                        <li className="mb-2">
-                            <a href="#" className="dark:hover:text-violet-700">Brand Center</a>
-                        </li>
-                        <li className="mb-2">
                             <a href="#" className="dark:hover:text-violet-700">Blog</a>
                         </li>
                     </ul>
@@ -24,12 +18,6 @@ function Footer() {
                 <div>
                     <h2 className="mb-3 text-sm font-semibold text-violet-300 uppercase">Help center</h2>
                     <ul className="text-violet-300">
-                        <li className="mb-2">
-                            <a href="#" className="dark:hover:text-violet-700">Discord Server</a>
-                        </li>
-                        <li className="mb-2">
-                            <a href="#" className="dark:hover:text-violet-700">Twitter</a>
-                        </li>
                         <li className="mb-2">
                             <a href="#" className="dark:hover:text-violet-700">Facebook</a>
                         </li>
@@ -45,9 +33,6 @@ function Footer() {
                             <a href="#" className="dark:hover:text-violet-700">Privacy Policy</a>
                         </li>
                         <li className="mb-2">
-                            <a href="#" className="dark:hover:text-violet-700">Licensing</a>
-                        </li>
-                        <li className="mb-2">
                             <a href="#" className="dark:hover:text-violet-700">Terms &amp; Conditions</a>
                         </li>
                     </ul>
@@ -60,12 +45,6 @@ function Footer() {
                         </li>
                         <li className="mb-2">
                             <a href="#" className="dark:hover:text-violet-700">Android</a>
-                        </li>
-                        <li className="mb-2">
-                            <a href="#" className="dark:hover:text-violet-700">Windows</a>
-                        </li>
-                        <li className="mb-2">
-                            <a href="#" className="dark:hover:text-violet-700 ">MacOS</a>
                         </li>
                     </ul>
                 </div>
