--- src/Composants/Searchbutton.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Searchbutton() {
-  return (  
-      <a href="/SearchPage">
-      <button type="submit" className="mb-6 text-white bg-violet-400 transition-all duration-1000 hover:bg-violet-700 focus:outline-none font-medium rounded-lg text-sm sm:w-auto m-auto px-5 py-2.5 text-center my-2">Recherche</button>
-      </a>        
-  )
-}
-
-export default Searchbutton;
